+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 v1.8.2
+- 移除```wrapUpdateSQL```,```wrapInsertSliceSQL```复写函数
+- ```WrapUpdateStructFinder```修改为```WrapUpdateSQLValue```变量
 - 重构更新逻辑,使用```entityStructCache```替代分散的map
 - 完善文档,注释
 
@@ -120,7 +122,7 @@ v1.6.1
  - 使用RegisterCustomDriverValueConver函数替代CustomDriverValueMap变量,将 ```zorm.CustomDriverValueMap["*dm.DmClob"] = CustomDMText{}```修改为```zorm.RegisterCustomDriverValueConver("TEXT", CustomDMText{})```,达梦数据库重新复制示例代码,重新复制!!重新复制!!!
  - 重写sqlRowsValues函数,支持查询单个字段,Struct类型接收
  - 简化自增序列的实现,使用string代替map[string]string
- - 使用OverrideFunc重写zorm的函数,暴露WrapUpdateStructFinder函数
+ - 使用OverrideFunc重写zorm的函数,暴露WrapUpdateSQLValue函数
  - 去掉kingbase列的大写转换,修改字符串拼接方式,提升性能
  - BindContextDisableTransaction 用在不使用事务更新数据库的场景,强烈建议不要使用这个方法,更新数据库必须有事务!!!
  - 增加查询没有返回列的判断,特殊情况可以使用Query执行更新语句,绕过事务检查(不建议)